# src/main/resources/api/dev-practice.openapi.yaml
openapi: 3.0.3
info:
  title: La Forge API
  version: 0.1.0
  description: |
    Contrat du prototype de laboratoire d’entraînement technique. Texte métier en français, identifiants API en anglais.

    Périmètre : profil, catalogue, exercices versionnés, tentatives, indices, corrigés, révisions, veille, défis collectifs, édition et tuteur IA optionnel. Pas d’exécution distante de code, génération automatique de contenus, classement RH ou gestion multi-entreprises. Une installation correspond à une entreprise ou un groupe pilote. Les rôles sont gérés hors API.

    Authentification : compte créé localement via /auth/registrations (email + mot de passe), activé par confirmation
    du jeton de vérification reçu par courriel (/auth/verifications). Connexion via /auth/sessions : le serveur émet
    lui-même un bearer JWT de courte durée (signature RS256) accompagné d’un jeton de renouvellement opaque
    (/auth/sessions/refresh). Le serveur valide signature, issuer, audience et expiration ; utilisateur courant déduit
    de sub, jamais du corps de requête. Rôles LEARNER, AUTHOR, REVIEWER, ADMIN cumulatifs ; tous les comptes peuvent
    pratiquer dès activation. x-required-roles signifie au moins un rôle de la liste, plus les contraintes de propriété
    décrites. ADMIN n’accorde pas l’accès aux tentatives privées des autres.

...

...

...

...

...

...

...
servers:
- url: /api/v1
  description: Même origine que le frontend ; préfixe de la v1.
security:
- bearerAuth: []
...
paths:
...
  /attempts:
    post:
      tags:
      - Practice
      operationId: createAttempt
      summary: Commencer une tentative
      description: 'Utilisateur courant uniquement. En cas de tentative IN_PROGRESS existante hors défi : 409. Pour
        un défi, exige inscription, état OPEN, absence de tentative précédente et correspondance exercice/version
        ; toute pratique antérieure soumise ou indice déjà révélé sur cette version hors de ce défi interdit la
        tentative de défi (409). Pour reviewItemId, exige révision due et aucune tentative liée en cours ; soumettre
        clôture cette révision. Le statut initial est IN_PROGRESS.'
      responses:
        '201':
          description: Commencer une tentative.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attempt'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAttempt'
  /me/attempts:
    get:
      tags:
      - Practice
      operationId: listMyAttempts
      summary: Lister mes tentatives
      description: Ordre startedAt décroissant puis id. Ne retourne que les tentatives de l’appelant.
      responses:
        '200':
          description: Lister mes tentatives.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttemptPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - name: status
        in: query
        schema:
          type: string
          enum:
          - IN_PROGRESS
          - SUBMITTED
          - ABANDONED
        description: ''
  /attempts/{attemptId}:
    get:
      tags:
      - Practice
      operationId: getAttempt
      summary: Lire ma tentative
      description: Accès propriétaire uniquement, sinon 404.
      responses:
        '200':
          description: Lire ma tentative.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attempt'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: attemptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /attempts/{attemptId}/submission:
    put:
      tags:
      - Practice
      operationId: submitAttempt
      summary: Soumettre la réponse définitive
      description: 'Transition atomique IN_PROGRESS → SUBMITTED. Valide le format et les références ; réponse ensuite
        immuable. Même réponse canonique déjà soumise : 200 sans répéter les effets ; réponse différente : 409.
        Défi fermé et tentative non soumise : 409. Débloque le débrief ; déclenche une révision si choix incorrect
        ou confiance ≤ 2. Clôture la révision source le cas échéant. Aucune dépendance à l’IA.'
      responses:
        '200':
          description: Soumettre la réponse définitive.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attempt'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: attemptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAnswer'
  /attempts/{attemptId}/abandon:
    post:
      tags:
      - Practice
      operationId: abandonAttempt
      summary: Abandonner une tentative individuelle
      description: 'IN_PROGRESS → ABANDONED, appel répété : 200. Tentative soumise ou liée à un défi : 409. Une
        révision source reste due. Propriétaire uniquement.'
      responses:
        '200':
          description: Abandonner une tentative individuelle.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attempt'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: attemptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /attempts/{attemptId}/hints:
    post:
      tags:
      - Practice
      operationId: revealHint
      summary: Débloquer un indice
      description: 'Propriétaire, tentative IN_PROGRESS. Les indices se débloquent dans l’ordre à partir de 1. Un
        niveau déjà révélé retourne le même indice sans nouvel effet ; saut de niveau : 409 ; niveau inexistant
        : 422. Indices interdits en défi (403) pour conserver la comparaison des raisonnements initiaux.'
      responses:
        '200':
          description: Débloquer un indice.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hint'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: attemptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HintRequest'
  /attempts/{attemptId}/debrief:
    get:
      tags:
      - Practice
      operationId: getDebrief
      summary: Lire le corrigé et le résultat
      description: Propriétaire uniquement. SUBMITTED requis, sinon 409. Correspond exactement à la version figée
        dans la tentative. Aucun appel au modèle nécessaire.
      responses:
        '200':
          description: Lire le corrigé et le résultat.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Debrief'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: attemptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /attempts/{attemptId}/self-assessment:
    put:
      tags:
      - Practice
      operationId: setSelfAssessment
      summary: Enregistrer mon autoévaluation
      description: SUBMITTED requis. Crée ou actualise un unique ReviewItem de suivi pour cette tentative. AGAIN/HARD/GOOD/EASY
        planifient respectivement à submittedAt + 1/3/7/21 jours (intervalles fixes du pilote, 24 h). Une date passée
        reste due immédiatement. Répéter la même autoévaluation ne décale pas la date. Une révision déjà commencée
        ou terminée ne peut plus être replanifiée (409). Remplace la planification automatique initiale.
      responses:
        '200':
          description: Enregistrer mon autoévaluation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: attemptId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SelfAssessment'
  /me/reviews:
    get:
      tags:
      - Practice
      operationId: listMyReviews
      summary: Lister mes révisions
      description: 'Par défaut DUE. Ordre dueAt croissant puis id. Révision automatique : submittedAt + 1 jour pour
        erreur objective, sinon + 3 jours si confiance ≤ 2 ; une seule entrée par tentative. Même version rejouée
        dans le MVP ; variantes éditoriales possibles ultérieurement. État DUE calculé par rapport à l’heure serveur.'
      responses:
        '200':
          description: Lister mes révisions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - name: state
        in: query
        schema:
          type: string
          enum:
          - DUE
          - SCHEDULED
          - COMPLETED
        description: ''
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Jeton du fournisseur d’identité configuré. Acquisition hors périmètre.
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Nombre maximal d’éléments.
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        maxLength: 1000
      description: Curseur opaque retourné par la page précédente.
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: 'UUID généré par le client, unique pour utilisateur + méthode + chemin. Conservation 24 h. Même
        clé et même corps : réponse originale sans répéter l’effet. Autre corps : 409. Déduplication atomique, y
        compris des appels IA. Opération encore en cours : 409 INVALID_STATE, réessayer avec la même clé.'
  responses:
    BadRequest:
      description: Requête syntaxiquement invalide.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthenticated:
      description: Jeton absent, expiré ou invalide.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        WWW-Authenticate:
          description: Challenge Bearer.
          schema:
            type: string
          example: Bearer
    Forbidden:
      description: Rôle insuffisant ou déblocage pédagogique non acquis.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Ressource inexistante ou non accessible dans le périmètre de l’appelant.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Transition interdite, tentative déjà finalisée, version de base obsolète ou conflit d’idempotence.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ValidationFailed:
      description: Données ou contraintes métier invalides ; violations détaille les champs.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    TooManyRequests:
      description: Quota dépassé ; respecter Retry-After.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        Retry-After:
          description: Délai minimal en secondes avant nouvel essai, si disponible.
          schema:
            type: integer
            format: int32
            minimum: 0
    InternalError:
      description: Erreur serveur inattendue.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      properties:
        type:
          type: string
          format: uri-reference
          example: urn:dev-practice:problem:invalid-state
        title:
          type: string
          maxLength: 200
          minLength: 1
        status:
          type: integer
          format: int32
          minimum: 400
          maximum: 599
        detail:
          type: string
          maxLength: 4000
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          enum:
          - BAD_REQUEST
          - UNAUTHENTICATED
          - FORBIDDEN
          - NOT_FOUND
          - INVALID_STATE
          - IDEMPOTENCY_CONFLICT
          - STALE_VERSION
          - VALIDATION_FAILED
          - PRECONDITION_REQUIRED
          - RATE_LIMITED
          - AI_UNAVAILABLE
          - INTERNAL_ERROR
        traceId:
          type: string
          maxLength: 100
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
          maxItems: 100
      required:
      - type
      - title
      - status
      - code
      example:
        type: urn:dev-practice:problem:invalid-state
        title: Tentative déjà soumise
        status: 409
        code: INVALID_STATE
        detail: La réponse définitive ne peut plus être modifiée.
    Violation:
      type: object
      additionalProperties: false
      properties:
        field:
          type: string
          maxLength: 200
          minLength: 1
        message:
          type: string
          maxLength: 4000
      required:
      - field
      - message
...
    ExerciseSummary:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
          format: int32
          minimum: 1
        title:
          type: string
          maxLength: 200
          minLength: 1
        type:
          $ref: '#/components/schemas/ExerciseType'
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        topicIds:
          type: array
          items:
            type: string
            format: uuid
          maxItems: 20
          uniqueItems: true
        estimatedMinutes:
          type: integer
          format: int32
          minimum: 1
          maximum: 60
        technologies:
          type: array
          items:
            $ref: '#/components/schemas/TechnologyRequirement'
          maxItems: 10
        publishedAt:
          type: string
          format: date-time
      required:
      - id
      - version
      - title
      - type
      - difficulty
      - topicIds
      - estimatedMinutes
      - technologies
      - publishedAt
...
    RubricCriterion:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          maxLength: 50
          minLength: 1
        label:
          type: string
          maxLength: 200
          minLength: 1
        importance:
          type: string
          enum:
          - BLOCKING
          - MAJOR
          - MINOR
        explanationMarkdown:
          type: string
          maxLength: 20000
          minLength: 1
      required:
      - id
      - label
      - importance
      - explanationMarkdown
    Correction:
      type: object
      additionalProperties: false
      properties:
        explanationMarkdown:
          type: string
          maxLength: 20000
          minLength: 1
        acceptedAnswers:
          type: array
          items:
            type: string
            maxLength: 4000
            minLength: 1
          maxItems: 20
        correctChoiceIds:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 10
          uniqueItems: true
        expectedVerdicts:
          type: array
          items:
            type: string
            enum:
            - APPROVE
            - REQUEST_CHANGES
            - NEED_INFORMATION
          maxItems: 3
          uniqueItems: true
        criteria:
          type: array
          items:
            $ref: '#/components/schemas/RubricCriterion'
          maxItems: 20
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
          maxItems: 20
          minItems: 1
      required:
      - explanationMarkdown
      - acceptedAnswers
      - criteria
      - sources
      description: Pour un choix, correctChoiceIds est obligatoire et référence les choix proposés (exactement un
        pour SINGLE_CHOICE). Pour REVIEW, expectedVerdicts est obligatoire et non vide. Les propriétés incompatibles
        sont interdites. Une réponse libre ne reçoit pas de note automatique de qualité.
    Hint:
      type: object
      additionalProperties: false
      properties:
        level:
          type: integer
          format: int32
          minimum: 1
          maximum: 5
        markdown:
          type: string
          maxLength: 20000
          minLength: 1
      required:
      - level
      - markdown
...
    FreeTextAnswer:
      type: object
      additionalProperties: false
      properties:
        kind:
          type: string
          enum:
          - FREE_TEXT
        text:
          type: string
          maxLength: 20000
          minLength: 1
        reasoning:
          type: string
          maxLength: 10000
          minLength: 1
        confidence:
          type: integer
          format: int32
          minimum: 1
          maximum: 5
      required:
      - kind
      - text
      - reasoning
      - confidence
    SingleChoiceAnswer:
      type: object
      additionalProperties: false
      properties:
        kind:
          type: string
          enum:
          - SINGLE_CHOICE
        choiceId:
          type: string
          maxLength: 50
          minLength: 1
        reasoning:
          type: string
          maxLength: 10000
          minLength: 1
        confidence:
          type: integer
          format: int32
          minimum: 1
          maximum: 5
      required:
      - kind
      - choiceId
      - reasoning
      - confidence
      example:
        kind: SINGLE_CHOICE
        choiceId: immutable-copy
        reasoning: Je distingue copie indépendante et simple vue non modifiable.
        confidence: 4
    MultipleChoiceAnswer:
      type: object
      additionalProperties: false
      properties:
        kind:
          type: string
          enum:
          - MULTIPLE_CHOICE
        choiceIds:
          type: array
          items:
            type: string
            maxLength: 50
            minLength: 1
          maxItems: 10
          minItems: 1
          uniqueItems: true
        reasoning:
          type: string
          maxLength: 10000
          minLength: 1
        confidence:
          type: integer
          format: int32
          minimum: 1
          maximum: 5
      required:
      - kind
      - choiceIds
      - reasoning
      - confidence
...
    Answer:
      oneOf:
      - $ref: '#/components/schemas/FreeTextAnswer'
      - $ref: '#/components/schemas/SingleChoiceAnswer'
      - $ref: '#/components/schemas/MultipleChoiceAnswer'
      - $ref: '#/components/schemas/ReviewAnswer'
      discriminator:
        propertyName: kind
        mapping:
          FREE_TEXT: '#/components/schemas/FreeTextAnswer'
          SINGLE_CHOICE: '#/components/schemas/SingleChoiceAnswer'
          MULTIPLE_CHOICE: '#/components/schemas/MultipleChoiceAnswer'
          REVIEW: '#/components/schemas/ReviewAnswer'
      description: Le kind doit correspondre au responseSpec de la version tentée. Choix, fichiers et lignes sont
        validés côté serveur (422).
    CreateAttempt:
      type: object
      additionalProperties: false
      properties:
        exerciseId:
          type: string
          format: uuid
        exerciseVersion:
          type: integer
          format: int32
          minimum: 1
        challengeId:
          type: string
          format: uuid
        reviewItemId:
          type: string
          format: uuid
      required:
      - exerciseId
      - exerciseVersion
      description: Version explicite pour éviter de changer d’énoncé en cours de parcours. challengeId et reviewItemId
        sont mutuellement exclusifs. Le serveur vérifie la correspondance exercice/version et la propriété de la
        révision.
      example:
        exerciseId: a3a4bd63-d3de-4875-9930-1027edb2a670
        exerciseVersion: 1
    Attempt:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        exerciseId:
          type: string
          format: uuid
        exerciseVersion:
          type: integer
          format: int32
          minimum: 1
        challengeId:
          type: string
          format: uuid
        reviewItemId:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - IN_PROGRESS
          - SUBMITTED
          - ABANDONED
        startedAt:
          type: string
          format: date-time
        submittedAt:
          type: string
          format: date-time
        answer:
          $ref: '#/components/schemas/Answer'
        revealedHints:
          type: array
          items:
            $ref: '#/components/schemas/Hint'
          maxItems: 5
        debriefAvailable:
          type: boolean
      required:
      - id
      - exerciseId
      - exerciseVersion
      - status
      - startedAt
      - revealedHints
      - debriefAvailable
      description: Privé au propriétaire, y compris vis-à-vis des administrateurs dans cette API. answer absent
        avant soumission. Une seule tentative IN_PROGRESS par utilisateur, exercice et version hors défi. Une seule
        tentative par participant et défi, non abandonnable.
    SubmitAnswer:
      type: object
      additionalProperties: false
      properties:
        answer:
          $ref: '#/components/schemas/Answer'
      required:
      - answer
    HintRequest:
      type: object
      additionalProperties: false
      properties:
        level:
          type: integer
          format: int32
          minimum: 1
          maximum: 5
      required:
      - level
    Debrief:
      type: object
      additionalProperties: false
      properties:
        attemptId:
          type: string
          format: uuid
        exerciseVersion:
          type: integer
          format: int32
          minimum: 1
        correction:
          $ref: '#/components/schemas/Correction'
        objectiveResult:
          type: string
          enum:
          - CORRECT
          - INCORRECT
          - NOT_AUTO_GRADED
        usedHintCount:
          type: integer
          format: int32
          minimum: 0
          maximum: 5
        selfAssessment:
          $ref: '#/components/schemas/SelfAssessment'
      required:
      - attemptId
      - exerciseVersion
      - correction
      - objectiveResult
      - usedHintCount
      description: 'Résultat déterministe seulement pour les choix : égalité exacte avec l’ensemble attendu. Les
        réponses libres et revues sont NOT_AUTO_GRADED. Le corrigé est disponible immédiatement après soumission,
        même sans IA.'
    SelfAssessment:
      type: object
      additionalProperties: false
      properties:
        mastery:
          type: string
          enum:
          - AGAIN
          - HARD
          - GOOD
          - EASY
        note:
          type: string
          maxLength: 2000
      required:
      - mastery
      example:
        mastery: HARD
        note: Je comprends le corrigé mais veux refaire un cas similaire.
    ReviewItem:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        exercise:
          $ref: '#/components/schemas/ExerciseSummary'
        sourceAttemptId:
          type: string
          format: uuid
        dueAt:
          type: string
          format: date-time
        reason:
          type: string
          enum:
          - INCORRECT
          - LOW_CONFIDENCE
          - SELF_ASSESSMENT
        state:
          type: string
          enum:
          - DUE
          - SCHEDULED
          - COMPLETED
      required:
      - id
      - exercise
      - sourceAttemptId
      - dueAt
      - reason
      - state
...
    AttemptPage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Attempt'
          maxItems: 100
        nextCursor:
          type: string
          maxLength: 1000
      required:
      - items
      description: Pagination par curseur opaque ; nextCursor absent à la dernière page. Aucun total garanti. Un
        curseur doit être réutilisé avec les mêmes filtres.
    ReviewPage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReviewItem'
          maxItems: 100
        nextCursor:
          type: string
          maxLength: 1000
      required:
      - items
      description: Pagination par curseur opaque ; nextCursor absent à la dernière page. Aucun total garanti. Un
        curseur doit être réutilisé avec les mêmes filtres.
...